+++ anyblok_marshmallow/__init__.py
@@ -1,6 +1,7 @@
 # This file is a part of the AnyBlok / Marshmallow api project
 #
 #    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -8,3 +9,4 @@
 
 from .schema import ModelSchema  # noqa
 from .exceptions import RegistryNotFound  # noqa
+from .fields import Nested, File  # noqa