+++ anyblok_marshmallow/tests/test_fields.py
@@ -1,6 +1,7 @@
 # This file is a part of the AnyBlok / Marshmallow project
 #
 #    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -9,7 +10,7 @@ from anyblok.tests.testcase import DBTestCase
 from anyblok_marshmallow import ModelSchema
 from anyblok_marshmallow.fields import Nested, File
 from anyblok import Declarations
-from anyblok.column import Integer, LargeBinary
+from anyblok.column import Integer, LargeBinary, Selection
 from anyblok.field import Function
 from anyblok.relationship import One2One
 from os import urandom
@@ -45,6 +46,25 @@ class TestField(DBTestCase):
         class Exemple:
             id = Integer(primary_key=True)
             file = LargeBinary()
+
+    def add_field_selection_with_object(self):
+
+        @Declarations.register(Declarations.Model)
+        class Exemple:
+            id = Integer(primary_key=True)
+            name = Selection(selections={'foo': 'bar', 'bar': 'foo'},
+                             default='foo')
+
+    def add_field_selection_with_classmethod(self):
+
+        @Declarations.register(Declarations.Model)
+        class Exemple:
+            id = Integer(primary_key=True)
+            name = Selection(selections='get_modes', default='foo')
+
+            @classmethod
+            def get_modes(cls):
+                return {'foo': 'bar', 'bar': 'foo'}
 
     def test_dump_function(self):
         registry = self.init_registry(self.add_field_function)
@@ -84,6 +104,128 @@ class TestField(DBTestCase):
             registry=registry, context={'model': "Model.Exemple"})
         errors = exemple_schema.validate(dump_data)
         self.assertFalse(errors)
+
+    def test_dump_selection_with_object(self):
+        registry = self.init_registry(self.add_field_selection_with_object)
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+
+        exemple = registry.Exemple.insert()
+        data, errors = exemple_schema.dump(exemple)
+        self.assertFalse(errors)
+        self.assertEqual(
+            data,
+            {
+                'id': exemple.id,
+                'name': 'foo',
+            }
+        )
+
+    def test_load_selection_with_object_ok(self):
+        registry = self.init_registry(self.add_field_selection_with_object)
+        dump_data = {
+            'id': 1,
+            'name': 'foo',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        data, errors = exemple_schema.load(dump_data)
+        self.assertEqual(data, dump_data)
+        self.assertFalse(errors)
+
+    def test_load_selection_with_object_ko(self):
+        registry = self.init_registry(self.add_field_selection_with_object)
+        dump_data = {
+            'id': 1,
+            'name': 'wrong_value',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        data, errors = exemple_schema.load(dump_data)
+        self.assertEqual(errors, {'name': ['Not a valid choice.']})
+
+    def test_validate_selection_with_object_ok(self):
+        registry = self.init_registry(self.add_field_selection_with_object)
+        dump_data = {
+            'id': 1,
+            'name': 'foo',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        errors = exemple_schema.validate(dump_data)
+        self.assertFalse(errors)
+
+    def test_validate_selection_with_object_ko(self):
+        registry = self.init_registry(self.add_field_selection_with_object)
+        dump_data = {
+            'id': 1,
+            'name': 'wrong_value',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        errors = exemple_schema.validate(dump_data)
+        self.assertEqual(errors, {'name': ['Not a valid choice.']})
+
+    def test_dump_selection_with_classmethod(self):
+        registry = self.init_registry(self.add_field_selection_with_classmethod)
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+
+        exemple = registry.Exemple.insert()
+        data, errors = exemple_schema.dump(exemple)
+        self.assertFalse(errors)
+        self.assertEqual(
+            data,
+            {
+                'id': exemple.id,
+                'name': 'foo',
+            }
+        )
+
+    def test_load_selection_with_classmethod_ok(self):
+        registry = self.init_registry(self.add_field_selection_with_classmethod)
+        dump_data = {
+            'id': 1,
+            'name': 'foo',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        data, errors = exemple_schema.load(dump_data)
+        self.assertEqual(data, dump_data)
+        self.assertFalse(errors)
+
+    def test_load_selection_with_classmethod_ko(self):
+        registry = self.init_registry(self.add_field_selection_with_classmethod)
+        dump_data = {
+            'id': 1,
+            'name': 'wrong_value',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        data, errors = exemple_schema.load(dump_data)
+        self.assertEqual(errors, {'name': ['Not a valid choice.']})
+
+    def test_validate_selection_with_classmethod_ok(self):
+        registry = self.init_registry(self.add_field_selection_with_classmethod)
+        dump_data = {
+            'id': 1,
+            'name': 'foo',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        errors = exemple_schema.validate(dump_data)
+        self.assertFalse(errors)
+
+    def test_validate_selection_with_classmethod_ko(self):
+        registry = self.init_registry(self.add_field_selection_with_classmethod)
+        dump_data = {
+            'id': 1,
+            'name': 'wrong_value',
+        }
+        exemple_schema = ModelSchema(
+            registry=registry, context={'model': "Model.Exemple"})
+        errors = exemple_schema.validate(dump_data)
+        self.assertEqual(errors, {'name': ['Not a valid choice.']})
 
     def getExempleSchema(self):